third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/unloading-documents/unload/unload-main-frame-same-origin.window-expected.txt

This is a testharness.js-based test.
[FAIL] Unload runs in main frame when navigating same-origin.
  assert_equals: expected "ran" but got "not yet"
Harness: the test ran to completion.
